# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.AppliedFilters:
    properties:
      city:
        type: string
      country:
        type: string
      cuisine_type:
        type: string
      has_email:
        type: boolean
      has_phone:
        type: boolean
      has_social_media:
        type: boolean
      has_website:
        type: boolean
      industry:
        type: string
      specialties:
        items:
          type: string
        type: array
      sport_type:
        type: string
      sub_niche:
        type: string
      tattoo_style:
        type: string
      verified:
        type: boolean
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  models.LeadListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.LeadResponse'
        type: array
      filters:
        $ref: '#/definitions/models.AppliedFilters'
      pagination:
        $ref: '#/definitions/models.PaginationInfo'
    type: object
  models.LeadPreviewResponse:
    properties:
      estimated_count:
        type: integer
      quality_score_avg:
        type: number
      verified_count:
        type: integer
      verified_pct:
        type: number
      with_email_count:
        type: integer
      with_email_pct:
        type: number
      with_phone_count:
        type: integer
      with_phone_pct:
        type: number
    type: object
  models.LeadResponse:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      cuisine_type:
        type: string
      email:
        type: string
      id:
        type: integer
      industry:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      quality_score:
        type: integer
      social_media:
        additionalProperties:
          type: string
        type: object
      specialties:
        items:
          type: string
        type: array
      sport_type:
        type: string
      sub_niche:
        type: string
      tattoo_style:
        type: string
      verified:
        type: boolean
      website:
        type: string
    type: object
...
  models.PaginationInfo:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: localhost:7890
info:
  contact:
    email: support@industrydb.io
    name: API Support
    url: https://industrydb.io/support
  description: Industry-specific business data API. Verified. Affordable.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://industrydb.io/terms
  title: IndustryDB API
  version: "1.0"
paths:
...
  /leads:
    get:
      consumes:
      - application/json
      description: Search leads with filters (industry, location, contact info). Requires
        authentication.
      parameters:
      - description: Industry filter (tattoo, beauty, gym, restaurant)
        in: query
        name: industry
        type: string
      - description: Country code (US, GB, ES, etc.)
        in: query
        name: country
        type: string
      - description: City name
        in: query
        name: city
        type: string
      - description: Filter by email presence
        in: query
        name: has_email
        type: boolean
      - description: Filter by phone presence
        in: query
        name: has_phone
        type: boolean
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Results per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Search results
          schema:
            $ref: '#/definitions/models.LeadListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Usage limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search for business leads
      tags:
      - Leads
  /leads/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve detailed information about a specific lead. Requires authentication.
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lead details
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Lead not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get lead by ID
      tags:
      - Leads
  /leads/preview:
    get:
      consumes:
      - application/json
      description: Get estimated count and statistics for a search without spending
        credits. Useful for seeing data availability before performing an actual search.
      parameters:
      - description: Industry filter (tattoo, beauty, gym, restaurant)
        in: query
        name: industry
        type: string
      - description: Country code (US, GB, ES, etc.)
        in: query
        name: country
        type: string
      - description: City name
        in: query
        name: city
        type: string
      - description: Filter by email presence
        in: query
        name: has_email
        type: boolean
      - description: Filter by phone presence
        in: query
        name: has_phone
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Preview statistics
          schema:
            $ref: '#/definitions/models.LeadPreviewResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Preview search results without charging credits
      tags:
      - Leads
...
securityDefinitions:
  ApiKeyAuth:
    description: API Key for programmatic access (Business tier only)
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"